requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# NLP and IR
nltk>=3.8
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# ML and Deep Learning
scikit-learn>=1.3.0
torch>=2.0.0

# Experiment tracking and evaluation
mlflow>=2.8.0

# LLM integration
openai>=1.0.0

# Scalability
faiss-cpu>=1.7.4

# Utilities
pyyaml>=6.0
tqdm>=4.66.0

# Web framework (for demo)
streamlit>=1.28.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
